Allow running against a non-local narrative

+++ test/unit/run_tests.py
@@ -12,7 +12,6 @@ import subprocess
 import sys
 import argparse
 import threading
-import time
 import os
 import signal
 
@@ -34,57 +33,63 @@ argparser.add_argument(
 )
 options = argparser.parse_args(sys.argv[1:])
 
-nb_command = [
-    "kbase-narrative",
-    "--no-browser",
-    '--NotebookApp.allow_origin="*"',
-    "--ip=127.0.0.1",
-    "--port={}".format(JUPYTER_PORT),
-]
-
-if not hasattr(sys, "real_prefix"):
-    nb_command[0] = "kbase-narrative"
-
-nb_server = subprocess.Popen(
-    nb_command, stderr=subprocess.STDOUT, stdout=subprocess.PIPE, preexec_fn=os.setsid
-)
-
-# wait for notebook server to start up
-while 1:
-    line = nb_server.stdout.readline().decode("utf-8").strip()
-    if not line:
-        continue
-    print(line)
-    if "The Jupyter Notebook is running at:" in line:
-        break
-    if "is already in use" in line:
-        os.killpg(os.getpgid(nb_server.pid), signal.SIGTERM)
-        # nb_server.terminate()
-        raise ValueError(
-            "The port {} was already taken, kill running notebook servers".format(
-                JUPYTER_PORT
-            )
-        )
+nb_server = None
 
 
-def readlines():
-    """Print the notebook server output."""
+def run_narrative():
+    nb_command = [
+        "kbase-narrative",
+        "--no-browser",
+        '--NotebookApp.allow_origin="*"',
+        "--ip=127.0.0.1",
+        "--port={}".format(JUPYTER_PORT),
+    ]
+
+    if not hasattr(sys, "real_prefix"):
+        nb_command[0] = "kbase-narrative"
+
+    nb_server = subprocess.Popen(
+        nb_command, stderr=subprocess.STDOUT, stdout=subprocess.PIPE, preexec_fn=os.setsid
+    )
+
+    # wait for notebook server to start up
     while 1:
         line = nb_server.stdout.readline().decode("utf-8").strip()
-        if line:
-            print(line)
+        if not line:
+            continue
+        print(line)
+        if "The Jupyter Notebook is running at:" in line:
+            break
+        if "is already in use" in line:
+            os.killpg(os.getpgid(nb_server.pid), signal.SIGTERM)
+            # nb_server.terminate()
+            raise ValueError(
+                "The port {} was already taken, kill running notebook servers".format(
+                    JUPYTER_PORT
+                )
+            )
 
+    def readlines():
+        """Print the notebook server output."""
+        while 1:
+            line = nb_server.stdout.readline().decode("utf-8").strip()
+            if line:
+                print(line)
 
-thread = threading.Thread(target=readlines)
-thread.setDaemon(True)
-thread.start()
+    thread = threading.Thread(target=readlines)
+    thread.setDaemon(True)
+    thread.start()
 
-print("Jupyter server started!")
+    print("Jupyter server started!")
+
+    return nb_server
+
 
 resp_unit = 0
 resp_integration = 0
 try:
     if options.unit:
+        nb_server = run_narrative()
         print("starting unit tests")
         try:
             resp_unit = subprocess.check_call(
@@ -93,9 +98,13 @@ try:
         except subprocess.CalledProcessError as e:
             resp_unit = e.returncode
     if options.integration:
-        base_url = f"http://localhost:{JUPYTER_PORT}"
         env = os.environ.copy()
-        env["BASE_URL"] = base_url
+        base_url = env.get('BASE_URL', None)
+        if base_url is None:
+            print(f'Starting local narrative on {JUPYTER_PORT}')
+            nb_server = run_narrative()
+            base_url = f"http://localhost:{JUPYTER_PORT}"
+            env["BASE_URL"] = base_url
         print("starting integration tests")
         try:
             resp_integration = subprocess.check_call(
@@ -110,9 +119,13 @@ try:
             )
         except subprocess.CalledProcessError as e:
             resp_integration = e.returncode
+except Exception as e:
+    print(f'Error! {str(e)}')
 finally:
-    print("Done running tests, killing server.")
-    os.killpg(os.getpgid(nb_server.pid), signal.SIGTERM)
+    print("Done running tests.")
+    if nb_server is not None:
+        print("Killing server.")
+        os.killpg(os.getpgid(nb_server.pid), signal.SIGTERM)
     if resp_unit != 0:
         print(f"Unit tests completed with code {resp_unit}")
     if resp_integration != 0:
